dice log on one line only
Added the option of having the dice log on one line only. 
When eedGlobal.diceLogOnOneLine is set to true, the dice log is not written each time dice of one colour are rolled: instead the dice are recorded and writen at the end of the script, just before the calculated dice results.